+++ genmonlib/custom_controller.py
@@ -1069,7 +1069,7 @@ class CustomController(GeneratorController):
             StartInfo["nominalfrequency"] = self.NominalFreq
             StartInfo["phase"] = self.Phase
             StartInfo["PowerGraph"] = self.PowerMeterIsSupported()
-            StartInfo["TempSensors"] = self.GetTempSensorNames()
+            StartInfo["Sensors"] = self.GetSensorNames()
             StartInfo["NominalBatteryVolts"] = self.NominalBatteryVolts
             StartInfo["FuelCalculation"] = self.FuelTankCalculationSupported()
             StartInfo["FuelSensor"] = self.FuelSensorSupported()