+++ genmonlib/generac_HPanel.py
@@ -2211,7 +2211,7 @@ class HPanel(GeneratorController):
             StartInfo["nominalfrequency"] = self.NominalFreq
             StartInfo["phase"] = self.Phase
             StartInfo["PowerGraph"] = self.PowerMeterIsSupported()
-            StartInfo["TempSensors"] = self.GetTempSensorNames()
+            StartInfo["Sensors"] = self.GetSensorNames()
             StartInfo["NominalBatteryVolts"] = self.NominalBatteryVolts
             StartInfo["FuelCalculation"] = self.FuelTankCalculationSupported()
             StartInfo["FuelSensor"] = self.FuelSensorSupported()