+++ genmonlib/generac_powerzone.py
@@ -1533,7 +1533,7 @@ class PowerZone(GeneratorController):
             StartInfo["nominalfrequency"] = self.NominalFreq
             StartInfo["phase"] = self.Phase
             StartInfo["PowerGraph"] = self.PowerMeterIsSupported()
-            StartInfo["TempSensors"] = self.GetTempSensorNames()
+            StartInfo["Sensors"] = self.GetSensorNames()
             StartInfo["NominalBatteryVolts"] = self.NominalBatteryVolts
             StartInfo["FuelCalculation"] = self.FuelTankCalculationSupported()
             StartInfo["FuelSensor"] = self.FuelSensorSupported()